Rename any type Remove sdk dependencies
